insta_tok/danie_lalatun/requirements.txt
tiktok-uploader>=1.0.0
requests>=2.25.1
selenium>=4.0.0
instagrapi>=2.1.0
python-dotenv>=1.0.0